move the visitor
